update server folder
