+++ src/server/db.ts
@@ -1,8 +1,9 @@
-// src/server/db/client.ts
 import { PrismaClient } from "@prisma/client";
-import { env } from "../../env/server.mjs";
+
+import { env } from "../env/server.mjs";
 
 declare global {
+  // eslint-disable-next-line no-var
   var prisma: PrismaClient | undefined;
 }
