--- src/server/trpc/router/index.ts
@@ -1,2 +1,0 @@
-export { authRouter } from "./auth";
-export { guestbookRouter } from "./guestbook";
